Split to a dedicated stats file
